requirements.txt
streamlit
crewai>=0.51.0
crewai-tools>=0.1.0
pydantic>=2.7
python-dotenv>=1.0.1